refactor(remote_push): send remote info, add params
After each `receiver` is created, it sends it's `remoteInfo` to the
remote info channel. Each `pusher` listens for `remoteInfo`'s and adds
them to their config. This is how the `pusher` knows who to push its
dataset to when it calls `lib.Publish`.

This sending of information happens very close to `ShareInfo`, where
each instance broadcasts it own `ActorInfo` to each other instance.
Potentially, `ShareInfo` could allow each testcase to define what data
it wants broadcast to each other node and allow us to specify how each
node handles that information. However, let's wait until we have a
better understanding of how the role of an `Actor` will the similar or
different in future test cases.

We also added `datasetSize` and `pushersPerReceiver` as int params to
the `remotes push` testcase:

`datasetSize` is the size in byte of the dataset you want to create to
be sent to each remote

`pushersPerReceiver` allows you to set the ratio of pushers to receivers
you want in your test. The lowest it can be is one to one. If the number
of pushers per receiver you set is higher then the number of instances,
then the test will error. Each pusher pushes to each receiver.